backend/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy>=2.0
psycopg2-binary

pydantic
pydantic-settings
python-dotenv

PyJWT

stripe>=8.0.0